// front-barbershop/front-barbershop/src/features/cliente/types/cliente.types.ts
export type ReservationStatus =
  | "PENDIENTE"
  | "CONFIRMADA"
  | "EN_CURSO"
  | "COMPLETADA"
  | "CANCELADA";

export interface ReservationService {
  id: number;
  name: string;
  duration: number;
  price: number;
}

export interface ReservationResponse {
  id: number;
  barber: string;
  client: string;
  services: ReservationService[];
  date: string;
  startTime: string;
  endTime: string;
  totalPrice: number;
  status: ReservationStatus;
}